requirements.txt
google-generativeai
paddlepaddle
paddleocr
Pillow
numpy
python-dotenv

